Clean up file includes
My hope was to lower the startup overhead, which failed, but in the
meantime I learned that
    require 'foo'
and
    require 'foo.rb'
might load 'foo' twice. This is an issue due to
    require_relative 'polytexnic/formats'
    require_relative 'polytexnic/utils'
which also show up as part of the file globbing. This commit makes
sure double reqruing by chomping the '.rb' before running `require`.