Solve a bunch of interrelated problems
I thought that frontmatter in MOBI was too wide, but it actually
turned out to be a much deeper problem. *Every* chapter was too
wide unless it had math, due to only math chapters getting a #book id
in the EPUB fragment. This in turn interacted with the EPUB's copy of
polytexnic.css, which was (incorrectly) stripped of #book so that
it would validate. (I don't know why it doesn't validate with #book
present.) But one of the #book ids was surviving the stripping, and
it was managing to (a) be valid EPUB and (b) constrain the width of
chapters with #book ids, i.e., the ones with math.

This commit omit the #book div for all chapter fragments, and properly
strips #book from the CSS file. The result is nice page widths for
every chapter.

[Delivers #60205608]