Update the listener for the Markdown/PolyTeX mix
This was much harder than it seems, as it required listening to *.tex,
which broke the server because of things generated in the
`generated_polytex` directory. After flailing around with Listener
for a while, I finally realized I could filter things in the `change`
block instead of at the listener level. (The key limitation is that
Listener can't listen to `foo/*.md` and `bar/*.tex`; it has to listen
either to all .md or all .tex files, at least AFAICT. Eventually, I
was able to hack around this restriction.)

[#60546516]