Remove generated files from template gitignore
Things like softcover.sty, softcover.css, are generated, so in
in principle they don't need to be versioned, but in practice they
(a) change rarely and (b) are useful to have out of the box when
you clone a repository. This is especially true of the custom styles
for code blocks, which get sucked out of softcover.sty; when that
file isn't in the repository, the first run of a PDF build uses
the wrong font size for code because the softcover.sty file hasn't
yet been copied from the template when the LaTeX-to-HTML conversion
happens. By putting it in the repo, we prevent this issue from
happening.

[Delivers #69120810]