Attempt an article ToC
I'm giving up for now because it seems like the EPUB standard won't
allow links in the ToC to anything other than files listed in the
spine, which would mean breaking the article into separate files
just to have the ToC work. But you can already get the ToC by
clicking on 'Frontmatter' and navigating from there, which is
probably good enough.