Bump versions
I've bumped to 1.2 to reflect the new polytexnic gem version
number, which in turn reflects the new ability to include HTML
inside LaTeX documents for propagation to the HTML/EPUB/MOBI output.

The main use for this is to include divs, spans, etc., to allow
things like labeling exercises with CSS ids and doing conditional
display based on class names (such as showing or not showing a block
depending on whether we're in a Learn Enough course or a regular
document).