Reuse menu headings in MOBI ToC
I ran into the problem of chapter names that included LaTeX and wanted to
run the text through the PolyTeXnic pipeline, only to realize this problem
had already been solved when making the menu headings for the local HTML
display. I.e., I reduced it to a previously solved problem. QED.

UPDATE: It turns out it wasn't *exactly* the same. EPUB won't allow HTML in
chapter names, so I had to add an HTML stripper to remove tags and such. But
still, it was *almost* the same solution.